requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==2.3.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Flask-Limiter==3.5.0
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==1.24.3
opencv-python==4.8.0.74
opencv-contrib-python==4.8.0.74
Pillow==10.0.0
pyotp==2.9.0
qrcode==8.0
SQLAlchemy==2.0.20
typing_extensions==4.12.2
Werkzeug==2.3.7
email-validator==2.0.0.post2
python-dotenv==1.0.0